Reorganize generateDates function to single file
