+++ client/utils/generateDates.js
@@ -27,4 +27,20 @@ const generateFirstWeek = (year, month) => {
   return firstWeek;
 };
 
-export default generateFirstWeek;
+const generateDates = (year, month) => {
+  const dates = [];
+  const firstWeek = generateFirstWeek(year, month);
+  dates.push(firstWeek);
+
+  for (let i = 0; i < 5; i += 1) {
+    const week = dates[dates.length - 1];
+    const nextWeek = week.map((date) => {
+      return moment(date).add(7, 'days');
+    });
+    dates.push(nextWeek);
+  }
+
+  return dates;
+};
+
+export default generateDates;
